More unit tests and standardization
